git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@276 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ ports/ARM7/chtypes.h
@@ -28,7 +28,7 @@
 #include <stdint.h>
 #endif
 
-typedef int8_t          bool_t;
+typedef int32_t         bool_t;
 typedef uint8_t         tmode_t;
 typedef uint8_t         tstate_t;
 typedef uint16_t        tid_t;
@@ -39,6 +39,6 @@ typedef uint32_t        eventmask_t;
 typedef uint32_t        systime_t;
 typedef int32_t         cnt_t;
 
-#define INLINE      inline
+#define INLINE inline
 
 #endif /* _CHTYPES_H_ */
